Changed to strings on UI. Added MVM and MF.
