+++ app/src/main/java/com/example/pool/ui/main/MainViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.pool.ui.main
+
+import androidx.lifecycle.ViewModel
+
+class MainViewModel : ViewModel() {
+    // TODO: Implement ViewModel
+}
