+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">Pool</string>
+    <string name="ChlorineHint">Chlorine</string>
+    <string name="PoolText">Your Pool</string>
 </resources>
